feat(cpr-physic-model): Add CPR physic document model structure
